libceph: mark redirected ops with new REDIRECTED flag
It will be useful in the future for ceph OSDs to know whether an op has
been resent based on a redirect.  This mirrors a recent change in
userland.

We set the KNOWN_REDIR flag unconditionally to so that the OSD knows
whether the REDIRECTED flag is authoritative or not.

Signed-off-by: Sage Weil <sage@inktank.com>