sched: Wrapper for checking task_struct::on_rq
Implement task_queued() and use it everywhere instead of on_rq check.
No functional changes.

The only exception is we do not use the wrapper in check_for_tasks(),
because it requires to export task_queued() in global header files.
Next patch in series would return it back, so it doesn't matter.

Cc: pjt@google.com
Cc: oleg@redhat.com
Cc: rostedt@goodmis.org
Cc: umgwanakikbuti@gmail.com
Cc: tim.c.chen@linux.intel.com
Cc: mingo@kernel.org
Cc: nicolas.pitre@linaro.org
Signed-off-by: Kirill Tkhai <ktkhai@parallels.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/20140730214242.27604.26201.stgit@localhost