sched/fair: Remove double_lock_balance() from active_load_balance_cpu_stop()
Bad situation:

double_lock_balance() drops busiest_rq lock. The busiest_rq is *busiest*,
and a lot of tasks and context switches there. We are dropping the lock
and waiting for it again.

Let's just detach the task and once finally unlock it!

Warning: this admits unlocked using of can_migrate_task(), throttled_lb_pair(),
and task_hot(). I've added lockdep asserts to point on this.

Cc: peterz@infradead.org
Cc: pjt@google.com
Cc: oleg@redhat.com
Cc: rostedt@goodmis.org
Cc: umgwanakikbuti@gmail.com
Cc: ktkhai@parallels.com
Cc: tim.c.chen@linux.intel.com
Cc: mingo@kernel.org
Cc: nicolas.pitre@linaro.org
Signed-off-by: Kirill Tkhai <ktkhai@parallels.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/20140730214302.27604.67247.stgit@localhost